requirements.txt
# Core dependencies
torch>=1.9.0
open-clip-torch>=2.20.0
Pillow>=8.0.0
numpy>=1.21.0
PyYAML>=6.0
scikit-image>=0.19.0

# API dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Remote API dependencies
openai>=0.27.0
requests>=2.25.0
